Moved deactive account card in a separate component
Fixes #180
