+++ frontend/course-hub-ui/src/css/profile.css
@@ -53,7 +53,6 @@
 }
 
 .profile-edit-card{
-    margin-bottom: 10px;
     width: 100%;
     height: 100%;
     box-shadow: 0 1px 1px 0 rgba(0,28,36,0.5);
@@ -61,15 +60,15 @@
 }
 
 .card-title{
-    background-color: #fafafa;
+    background-color: #f0f0f0;
     border-bottom: 1px solid #d6d6d6;
     color: inherit;
     text-align: left;
     font-weight: 700;
     font-size: 1.8rem;
     padding-left: 10px;
-    padding-top: 10px;
-    padding-bottom: 10px;
+    padding-top: 15px;
+    padding-bottom: 15px;
 }
 
 .profile-card-body{
@@ -80,7 +79,6 @@
     color: inherit;
     font-weight: 700;
     text-align: left;
-    color: #3a3a3a;
     font-size: 13px;
     padding: 10px;
 }
@@ -110,20 +108,18 @@
 
 .main-content{
     position:relative;
-    /*z-index:1;*/
-
+    width: 98%;
 }
 
 .profile-image-card{
-    padding-left: 10px;
     background-color: #4f5356;
-    margin-bottom: 10px;
     width: 90%;
     height: 100%;
+    box-shadow: 0 1px 1px 0 rgba(0,28,36,0.5);
 }
 
 .profile-upload{
-    padding-top: 30%;
+    padding-top: 10%;
     padding-right: 10px;
 }
